permuter-campaign: land VC71 last-mile recoveries
Permuter-driven MSVC 7.1 match raises, committed at >=90%:

- objects_fix_for_deleted_object 97.4→100.0
- attachments_delete 97.7→100.0
- hs_update 97.6→100.0
- network_game_server_get_client_machine 97.7→100.0
- cache_copy_compressed_alloc 98.0→100.0
- device_group_set_actual_value 97.8→100.0
- FUN_00180570+FUN_00180660 97.7→100.0
- FUN_00111170 88.6→96.6
- player_add_equipment 86.5→90.6
- actor_action_allow_cover_seeking 86.1→100.0
- ai_debug_idle_look_addprop 86.3→98.0
- unit_place 94.8→95.8
- scenario_location_potentially_visible 94.6→100.0
- set_file_location_volume_name 93.2→95.7
- FUN_00062410 93.9→95.5
- find_objects_from_point_vector 92.7→93.8
- lrar_cache_new 92.5→93.2
- collision_surface_area 91.0→95.0
- unit_impact_melee_damage 90.8→94.9
- FUN_0011beb0 90.2→92.0
- collision_surface_test_line2d 96.0→98.8

Original 20-commit tip: backup/permuter_campaign-pre-squash (4866e2c93).

+++ src/halo/devices/devices.c
@@ -308,14 +308,14 @@ void device_set_never_appears_locked(int object_index,
  * request, whose exact meaning is unproven). The iterator is advanced once
  * before the loop guard, giving the original's do/while shape.
  */
-void device_group_set_actual_value(int device_group_index, float value)
+void device_group_set_actual_value(short device_group_index, float value)
 {
   short index;
   char *device_group;
   int iterator[4];
   char *object;
 
-  index = (short)device_group_index;
+  index = device_group_index;
 
   if (value < *(float *)0x2533c0) {
     value = 0.0f;
